Reverted latest changes.
svn path=/trunk/; revision=3473

+++ reactos/lib/msvcrt/io/find.c
@@ -1,4 +1,7 @@
-#include <msvcrti.h>
+#include <windows.h>
+#include <msvcrt/io.h>
+#include <msvcrt/string.h>
+#include <msvcrt/internal/file.h>
 
 
 int _findclose(int handle)
@@ -55,7 +58,7 @@ int _findfirst(const char *_name, struct _finddata_t *result)
   return hFindFile;
 }
 
-long _findfirsti64(const char *_name, struct _finddatai64_t *result)
+int _findfirsti64(const char *_name, struct _finddatai64_t *result)
 {
   WIN32_FIND_DATAA FindFileData;
   char dir[MAX_PATH];
@@ -123,7 +126,7 @@ int _findnext(int handle, struct _finddata_t *result)
   return 0;
 }
 
-int _findnexti64(long handle, struct _finddatai64_t *result)
+int _findnexti64(int handle, struct _finddatai64_t *result)
 {
   WIN32_FIND_DATAA FindFileData;
 
@@ -145,7 +148,7 @@ int _findnexti64(long handle, struct _finddatai64_t *result)
   return 0;
 }
 
-long _wfindfirst(const wchar_t *_name, struct _wfinddata_t *result)
+int _wfindfirst(const wchar_t *_name, struct _wfinddata_t *result)
 {
   WIN32_FIND_DATAW FindFileData;
   wchar_t dir[MAX_PATH];
@@ -191,7 +194,7 @@ long _wfindfirst(const wchar_t *_name, struct _wfinddata_t *result)
   return hFindFile;
 }
 
-long _wfindfirsti64(const wchar_t *_name, struct _wfinddatai64_t *result)
+int _wfindfirsti64(const wchar_t *_name, struct _wfinddatai64_t *result)
 {
   WIN32_FIND_DATAW FindFileData;
   wchar_t dir[MAX_PATH];
@@ -238,7 +241,7 @@ long _wfindfirsti64(const wchar_t *_name, struct _wfinddatai64_t *result)
   return hFindFile;
 }
 
-int _wfindnext(long handle, struct _wfinddata_t *result)
+int _wfindnext(int handle, struct _wfinddata_t *result)
 {
   WIN32_FIND_DATAW FindFileData;
 
@@ -259,7 +262,7 @@ int _wfindnext(long handle, struct _wfinddata_t *result)
   return 0;
 }
 
-int _wfindnexti64(long handle, struct _wfinddatai64_t *result)
+int _wfindnexti64(int handle, struct _wfinddatai64_t *result)
 {
   WIN32_FIND_DATAW FindFileData;
